+++ bin/tsconfig.json
@@ -3,5 +3,5 @@
     "compilerOptions": {
         "resolveJsonModule": true
     },
-    "exclude": ["../src"]
+    "exclude": ["../src", "../tests"]
 }
