+++ tsconfig.json
@@ -1,4 +1,4 @@
 {
     "extends": "@tsconfig/node14/tsconfig.json",
-    "include": ["src", "bin"]
+    "include": ["src", "bin", "tests"]
 }
